VM > 1 Support Added
Changed the underlying data types from the binary VRs from one item to
an array of items. For example, the data type for FloatingPointSingle is
no longer float. It is float[]. This will break your code if you have
implemented a previous version of ED. You will have to update
appropriately. Added Multiplicity readers and writers to make sure all
data is accessed and written appropriately.